Rename filters to dj_filters
The idea is to distinguish mathesar.api.filters from
mathesar.api.viewsets.filters, since one is referring to Django filters
(hence dj_filters) and the other to the Mathesar filter abstraction.
