+++ mathesar/api/serializers/databases.py
@@ -2,7 +2,7 @@ from django.urls import reverse
 from rest_framework import serializers
 
 from mathesar.api.display_options import DISPLAY_OPTIONS_BY_TYPE_IDENTIFIER
-from mathesar.api.filters import FILTER_OPTIONS_BY_TYPE_IDENTIFIER
+from mathesar.api.dj_filters import FILTER_OPTIONS_BY_TYPE_IDENTIFIER
 from mathesar.models import Database
