+++ mathesar/api/viewsets/databases.py
@@ -5,7 +5,7 @@ from rest_framework.mixins import ListModelMixin, RetrieveModelMixin
 from rest_framework.response import Response
 
 from mathesar.models import Database
-from mathesar.api.filters import DatabaseFilter
+from mathesar.api.dj_filters import DatabaseFilter
 from mathesar.api.pagination import DefaultLimitOffsetPagination
 from mathesar.api.serializers.databases import DatabaseSerializer, TypeSerializer
