+++ mathesar/api/viewsets/tables.py
@@ -8,7 +8,7 @@ from rest_framework.response import Response
 from sqlalchemy.exc import ProgrammingError, DataError, IntegrityError
 
 from db.types.exceptions import UnsupportedTypeException
-from mathesar.api.filters import TableFilter
+from mathesar.api.dj_filters import TableFilter
 from mathesar.api.pagination import DefaultLimitOffsetPagination
 from mathesar.api.serializers.tables import TableSerializer, TablePreviewSerializer
 from mathesar.models import Table
